test: 오픈 서치 테스트 모듈 추가
- 불필요한 인증서 제거
- 웹 모듈에서 오픈 서치 테스트 모듈 사용하도록 변경
- 배치 모듈에서 오픈 서치 테스트 모듈 사용하도록 변경
