+++ batch/src/test/resources/application.yml
@@ -30,12 +30,9 @@ logging:
     root: info
 
 elastic:
-  username: ${ELASTIC_USERNAME:elastic}
-  password: ${ELASTIC_PASSWORD:changeit}
+  username: ${ELASTIC_USERNAME:admin}
+  password: ${ELASTIC_PASSWORD:admin}
   host: ${ELASTIC_HOST:localhost}
   port: ${ELASTIC_PORT:9205}
-  truststore:
-    location: ${ELASTIC_TRUSTSTORE_LOCATION:src/test/resources/test-keystore/test-ca.p12}
-    password: ${ELASTIC_TRUSTSTORE_PASSWORD:changeit}
   connection-timeout-millis: 30000
   socket-timeout-millis: 30000
