+++ opensearch-test/build.gradle
@@ -1,0 +1,20 @@
+plugins {
+    id 'java'
+    id 'org.springframework.boot'
+    id 'io.spring.dependency-management'
+
+    /* kotlin */
+    id 'org.jetbrains.kotlin.jvm'
+    id "org.jetbrains.kotlin.kapt"
+    id "org.jetbrains.kotlin.plugin.jpa"
+    id "org.jetbrains.kotlin.plugin.spring"
+}
+
+dependencies {
+    api("org.opensearch:opensearch-testcontainers:2.0.1")
+    api("org.testcontainers:junit-jupiter:${testContainersVersion}")
+}
+
+bootJar {
+    enabled = false
+}
